refactor!: move extension functions to their corresponding classes
BREAKING CHANGE: This changes the import paths for extension functions.
